[github] Add invoke task to run MacOS build Github actions and integrate it in Gitlab pipelines (#6384)
Adds invoke task to run a Github Action to build the MacOS Agent.
Adds Python classes to interact with the Github API as a Github App, in order to manipulate Github Actions workflows.
Adds jobs (on deploy pipelines) to build the signed MacOS Agent and deploy it to a staging bucket.

Co-authored-by: Pablo Baeyens <pablo.baeyens@datadoghq.com>
Co-authored-by: Albert Vaca Cintora <albert.vaca@datadoghq.com>