[gitlab] Implicitly declare pupernetes-dev dependencies (#6446)
Makes the pupernetes-dev dependencies implicit. Explicitly declaring a dependency on a manual job gives weird results (in this case, the job stays in a "created" state until the manual jobs are run, leaving our pipelines in a "Running" state).

On the other hand, that means people who want to run this job have to remember that they need to run the dependencies before running pupernetes-dev.