[orchestrator] ensure leader runs (#6592)
This pr ensures that for the orchestrator explorer the leader engine runs as long as the corresponding setting is enabled.
In case if the setting is not enabled but the orchestrator explorer a respective error will be shown.
Additionally the DCA output indicates whether the agent or follower is running.