[collector] Add `Cancel` to the `check.Check` interface (#6636)
1. Remove unused Collector.ReloadCheck
It was not used anywhere, and is probably broken.
Also, use a check.StubCheck to factorize test code on checks.

2. Add Cancel to the check.Check interface
Always called by the collector when a check is unscheduled. Allows
checks to stop and clean up background resources.
A timeout of 500ms is applied.
Not implemented in python checks for now. TODO: add python bindings

Co-authored-by: Ahmed Mezghani <38987709+ahmed-mez@users.noreply.github.com>