[DCA] [orchestrator] optimize scrubbing for orchestrator containers (#6619)
* [orchestrator] change orchestrator container scrubbing

This PR changes the default scrubbing algorithm we use for Containers (Env + Commands) from Regex search and replace to string.Contains(). Additionally, the new implementation has the possibility to add custom regex which will be run before the contains ones. This patch can be seen as a first step of multiple one to increase the performance of scrubbing.