[aggregator] Better support monotonic counts from openmetrics (#6663)
* [metrics] Simplify MonotonicCount implementation a little

Add tests, and fix behavior of `hasPreviousSample` so that a `0` is
always flushed when a sample lower than the sample of previous flush
is submitted.

* [metrics] Allow `MonotonicCount` to optionally flush the first value

When `FlushFirstValue` is enabled on a `MetricSample`, make the
aggregator flush the first value that's sampled for a (metric_name,
context).

This will be useful for the openmetrics check, where the assumptions
that make this first value correct can be verified by the check.

* [aggregator&rtloader] Support FlushFirstValue param on MonotonicCount

Added as an optional boolean parameter to `aggregator.submit_metric`.