Added `--flare` to `agent check` so checks can be reported in an agent flare (#6751)
In support cases it would be useful to record the results of an agent check so that it may be reported in a flare. Currently a workaround would be to redirect the output of the check command: datadog-agent check tcp_check > /var/log/datadog/check_cmd_tcp_check.log && datadog-agent flare. But it’s more of a hack than a supported behavior.

This PR introduces agent check <check_name> --flare which will write the results of the check to whatever the default logs directory is in a checks folder. For example: /var/log/datadog/checks.

The check logs are formatted as <check>_<checkID>_<timestamp>.log