Break down tagger telemetry per source and prefix
We suspect that there is an entity leak in the tagger, where some
entities never get deleted, even though the actual entities (as in,
pods, containers, tasks, etc) are long gone. To help in the
investigation, we want to find out which collector and which kind of
entity is leaking, so we're adding extra telemetry to point us in the
right direction.