pkg/trace/obfuscate: allow more context in top-level call SQL escape sequences (#6770)
Historically we've been fully obfuscating [SQL escape sequences](https://docs.oracle.com/cd/E13157_01/wlevs/docs30/jdbc_drivers/sqlescape.html). This can be problematic for some when their entire query is an escape sequence calling a procedure, such as:

    {[?=]call procedure-name[([parameter][,parameter]...)]}.

In such cases, the query would amount to `?` which is not very insightful and compacts many queries under the same ambiguous resource.

This change adds an exception such that when a query starts with a curly brace, it is not fully obfuscated as an escape sequence, but instead is treated like a regular query where only strings and numerics inside of it get obfuscated.

[Various databases and languages](http://media.datadirect.com/download/docs/slnk/devref/scalarfn.html) support these types of calls.