Enable transport for host metadata calls (#7042)
Before this change, the agent did not use the default/global transport used in the rest of the agent. As a result host metadata calls would not respect the no_proxy list when configured in the datadog.yaml.

This change uses the transport for all metadata calls as well as by default adds the metadata IP to the no_proxy list as to not break any existing users. This change also includes a new flag to disable this behavior and use the proxy for metadata calls.