[cmd/serverless] add support for Logs Collection in the Datadog Agent Extension (#6861)
* Add the Serverless Datadog Agent implementation.

See https://docs.datadoghq.com/serverless/datadog_lambda_library/extension/

* Fix 3rd party license list since new pkg are imported from the AWS SDK.

* forwarder: adapt the SyncDefaultForwarder to recent changes in the Forwarder interface.

* aggregator: fix unit tests with Flush now being a public function.

* serverless: linter compliance.

* serverless: errcheck compliance.

* serverless: unused code linter.

* serverless: linter compliance.

* serverless: add http server listening for server logs and sending logs to the logs agent

Introducing the `logs/input/channel` implementation supporting receiving logs through
a Go channel.

* serverless/logs: introduce a `NullAuditor` not doing anything with messages.

This has been introduced to not change the original pipeline process.
The original `Auditor` has been renamed `RegistryAuditor`.
`Auditor` is now an interface for both `RegistryAuditor` and `NullAuditor`.

* serverless: remove debug messages

* serverless/logs: add a sync `Flush()` method to the logs pipeline.

This commit is also adding the reading of the function ARN from the invoke event.

* serverless/logs: remove unwanted debug lines.

* serverless/logs: read REPORT messages and send AWS enhanced metrics using the DogStatsD server.

Currently supported enhanced metrics: billed_duration, duration, init_duration, max_memory_used.

* serverless/logs: enhanced metrics are distribution.

* serverless/logs: comment on the sync flush of logs then statsd.

* serverless/logs: configurable logs type we're subscribing to.

This commit also adds the parsing of the time in the AWS log.

* serverless/arn: remove version from the ARN string if any.

* serverless/logs: create aws package containing both AWS LogMessage and ARN support.

This way, we can use the timestamp from the AWS LogMesage while sending the logs
to the intake. This is what this commit is doing.

* serverless/logs: parse time up to the millisecond.

* serverless/logs: send the enhanced metrics with the proper timestamp.

* serverless/logs: comments and send the report log to the intake.

* serverless/logs: send aws.lambda.enhanced.memorysize

* serverless/logs: send formated platform logs.

* serverless/logs: cleaning comments here and there.

* config: fix log conflict while adding the serverless format.

* serverless/logs: linter compliance.

* cmd/serverless: start -> run

* dogstatsd: add an unit test in serverless mode.

* tasks: remove serverless test tag

* forwarder: rename `SyncDefaultForwarder` to `SyncForwarder` since it's not a default forwarder.

* [cmd/serverless] change default port for the HTTP server collecting logs

* serverless: support using DD_TAGS to add extra tags while sending metrics and logs.

For feature-parity with the Datadog Forwarder, we want to use this
configuration entry / environment variable to append tags to every
metric and log sent by the extension.

* serverless/logs: send the logs with the ARN and RequestId information.

* serverless: linter compliance

* serverless/logs: use the function name if available as the "service".

* serverless: hit the correct domain using the `AWS_LAMBDA_RUNTIME_API` env var.

* serverless: re-use the daemon http server to receive the logs from AWS.

Instead of spawning a new HTTP server (using yet another tcp port), we are re-using
the already spawned HTTP server which is already used to communicate with the libraries.

It also means there is no more configuration field to change the HTTP. This may have
to be addressed if 8124 is not available for some reasons.

* dogstatsd: fix a unit test.

* dogstatsd: fix a unit test.

* tasks/dogstatsd: slighlty increase the maximum valid size for the dogstatsd binary.

* misc: use the correct log-format for non-serverless logs.

This must be a merge error/artifact.