Update Android NDK and gomobile (#7206)
Bumps the Android NDK and updates the `init` command (it's no longer needs the NDK path).

It also bumps Gradle and the Gradle plugin versions, so they are compatible with the new NDK.

I also added `retrieve_linux_go_deps`: even if we still call inv deps, a lot of dependencies are reused this way.