Add telemetry around the new remote tagger machinery
We need to keep an eye on how the remote tagger is behaving, so we need
some extra telemetry:

* tagger.stream_errors tracks how often streams return an error and are
aborted. This is the main metric we should look at to make sure the
remote tagger actually works at all.

* tagger.events tracks how many events the tagger is sending out. This
is a good proxy metric for the network traffic we're generating.

* tagger.notifications tracks how many times the tagger has send a
message to a subscriber. More events but less notifications is good, too
many notifications with a small amount of events is bad.

* tagger.subscribers tracks how many subscribers there are. This should
be mostly static in the lifetime of an agent, and is here for
sanity-keeping.