[circleci] Disable PR-related checks on default branch (#7394)
Adds a check to the team_label, milestone and release_note job scripts to not run on the default branch.
Cleans up some unused code (our CircleCI settings only allow running jobs for PRs, tags or the default branch).
Updates the CircleCI config to not run these jobs on master.