[gdb image] Allow specifying the deb repo snapshot date (#7424)
Since the debian snapshot repos defined in comments in
/etc/apt/sources.list do not necessarily accurately reflect the deb
repos' state when the Agent image was built, add an option to allow
changing them so that the gdb image build is successful.

This is a temp fix: ideally the Agent image should keep track of when it
was built, so that we can re-use that date in this gdb image build to
use the snapshot of the deb repos for that date. Tracking this todo
in the README.

Also, make sure the install of gdb succeeds in fully non-interactive
mode.