[clcrunner] Allow runners to use different identifiers
We want to allow users to binpack Cluster Checks Runners to a single
node when necessary. Previously, each CLC runner was identifier by the
node name where they resided, so it was not possible to run more than
one per node without causing collisions.

This commit introduces a `clc_runner_id` config
option to allow users to override that now. It's probably not very
useful as a static value, so the intended usage is by injecting the
`DD_CLC_RUNNER_ID` env var in the pod spec as follows:

```
env:
- name: DD_CLC_RUNNER_ID
  valueFrom:
    fieldRef:
      fieldPath: metadata.name
```