[serverless] implement an adaptive flush automatically deciding the best moment to flush (#7232)
* serverless: compute the frequency of execution of the function.

In order to later select a flush strategy, this commit add the capability
to compute the frequency of execution of the function.

It is computed based on the 50 last invocations time.

* serverless/logs: support cancelling a flush.

Use `context.Context` to support cancelling the logs pipeline flush.

* serverless: auto-select a strategy for the best moment to flush.

* serverless: vet compliance.

* serverless: wrapping up + env var to disable the adaptive flush mechanism.

Add `DD_SERVERLESS_DISABLE_ADAPTIVE_FLUSH` to disable the adaptive flush mechanism
and always use the "flush at the end" strategy.

Also fix a bug in the sync flush of the logs agent.

* serverless: vlint compliance.

* serverless: address feedback

* serverless: linter compliance.

* serverless: remove the "flush at start" strategy which is already part of the periodic one.

* serverless: address feedback.

* Update pkg/serverless/interval.go

Co-authored-by: Darcy Rayner <50632605+DarcyRaynerDD@users.noreply.github.com>

Co-authored-by: Darcy Rayner <50632605+DarcyRaynerDD@users.noreply.github.com>