[orchestration] add cluster resource (#7551)
* [orchestrator] collect cluster resource for orchestrator explorer

This PR adds the ability for the orchestrator core check to additionally collect cluster resources. As defined in the payload. One might want to note that we will reuse the nodes informer to make sure to lessen the impact of the cluster resource. Additionally, this code makes sure to send the node resource even if cluster fails.