[build] Update CircleCI runner image to use Golang 1.15  (#7760)
* [build] Bump Circle CI builder Golang version to 1.15.10

We are upgrading our golang version so this change will be the first one
needed to get our code where it needs to be.

* [build] Cleaned up the CircleCI runner Dockerfile

The file had pretty unreadable apt dependencies so this commit cleans
those up to be more maintainable.

* [build] Update CircleCI builder to Golang 1.15.10

Since we're updating the Golang version, this change uses the
newly-pushed image in the building.