pkg/trace/agent: add "component2name" feature flag to facilitate Opentracing (#7690)
The "component2name" feature flag determines the span component tag to become
the span operation name.

It works around the incompatibility between Opentracing and Datadog where the
Opentracing operation name is many times invalid as a Datadog operation name (e.g. "/")
and in Datadog terms it's the resource. Here, we aim to make the component the
operation name to provide a better product experience.