[python loader tests] Add ability to instantiate mock `rtloader_*_t` structs (#7843)
Since Golang 1.15, we are no longer able to instantiate an opaque pointer to a
struct (`rtloader_*_t`) since it was an incomplete type. To ensure that we can
unit test our rtloader supporting logic that does a lot of sanity checks like
`somePtr != nil` we have to have a non-nil pointer that we can pass around (but
not directly use). By adding in helpers that can create such mock pointers, we
can ensure that tests using these opaque pointers are usable.