serverless: read a configuration file while running as a Lambda Extension. (#7783)
* serverless: do not start the HTTP server for expvar.

It is not used while running as a Serverless Agent.

* serverless: read a configuration file while running as a Lambda Extension.

The Datadog Lambda Extension will try to read the configuration
file in `/var/task/datadog.yaml`.
It will let us use some nice features such as a the logs scrubbing.

* serverless: document the logs filtering and scrubbing support.

* serverless: use the const to look for the configuration file instead of hardcoded value.

I've changed the const to look for the file into `/var/task` where it should belongs
on AWS Lambda. Note that the trace agent is also using this const.

* serverless: update comment indicating where is read the configuration

* serverless: remove logs scrubbing configuration samples.

We don't want the repository doc to be out of sync with public documentation.