[serializer] compress processes payload (#7911)
* [serializer] compress legacy resources payload

The only payload format still not compressed, and there is no reason not to.

* [serializer] rename method for clarity

SendJSONToV1Intake lost most of its functions over time, and is now used
only for legacy processes metadata. Rename the method to better reflect this.

Configuration setting is not renamed, as it would break user configuration.

Remove mention of V1 intake from the serializer's readme, because it no longer
provides a general purpose method for interacting with it.

* remove release note