[config] add `extra_tags` setting (#7872)
* [config] add new `extra_tags` setting

`extra_tags` specifies additional tags that will be appended to the host
tags. Allows tags to be set via both the configuration file and
environment variable at the same time.

* [config] deduplicate host tags

Addition of `extra_tags` introduces a chance that duplicate host tags
will be specified. Remove duplicates to reduce network load.

* [config] apply extra_tags everywhere

Refactor the code, so that the `extra_tags` setting is applied
everywhere, where `tags` was used.

* [config] add `extra_tags` to config template

* Update pkg/config/config_template.yaml

Co-authored-by: Lucas Pimentel-Ordyna <lucas.pimentel@datadoghq.com>

* [config] change how `extra_tags` is explained

Avoid potential confusion by only mentioning `extra_tags` in its
environment variable form, which is in line with the primary use case of
supplementing tags in the configuration file. But we still want users
that want to use it in the configuration file to be able to do so.

Co-authored-by: Lucas Pimentel-Ordyna <lucas.pimentel@datadoghq.com>