[logs/file/scanner] do not create a tailer if it is not concerning the container of the log source. (#8143)
* [logs/file/scanner] do not assign a tailer to a logfile if it is not concerning the container.

We also use the file scanner to look for containers and pods logs file, because of that
we have to make sure that the file we just detected is tagged with the correct
container ID.

The way k8s is storing files in /var/log/pods doesn't let us do that properly
(the filename doesn't contain the container ID).

However, the symlinks present in /var/log/containers are pointing to /var/log/pods files,
meaning that we can use them to validate that the file we have found is concerning us.

That's what the added function shouldIgnore is trying to do when the directory exists
and is readable.

This commit should fix the small logs duplication happening with short-lived containers.

See these links for more info:
   - https://github.com/kubernetes/kubernetes/issues/58638
   - https://github.com/fabric8io/fluent-plugin-kubernetes_metadata_filter/issues/105

* [logs/file/scanner] add a unit test validating the "shouldIgnore" while running with k8s.

* Add release note entry.

* FileInfo is an interface and could be nil.

* [file/input/scanner] validate the container ID parsed from the filename.

* Clean up after the test run.

* [file/input/scanner] use an early return in the shouldIgnore function.