Subscribe to rename docker events
This is to support the changes introduced in
4f76d13e62726bbcb0baa1dc4453260f262c4bf4, as without this the rename
events are filtered out and never reach the tagger.

Although this does affect other users of SubscribeToContainerEvents,
currently they'll handle this gracefully by ignoring the event
completely, or emitting a debug log. Since renames are a relatively rare
occurrence, I'm considering the extra log line acceptable.