Skip inspect cache when docker containers are renamed
This is a follow up to 4f76d13e62726bbcb0baa1dc4453260f262c4bf4. If the
cache is not bypassed, we may not update the `container_name` tag if the
previous container definition is still in the cache. The cache lives for
a very short period (currently 10 seconds), but the current use case for
this fix is a container that's renamed immediately after it's created,
therefore much earlier than the cache's expiration.