[rtloader] Enable dynamic interpreter binary path resolution (#8369)
* [rtloader] Add ability to set program name to rtloader

Previously, our RTLoader was unable to have a custom program name (via
`Py_SetProgramName`) which made use of libraries like `multiprocessing`
not work since they would try to spawn new processes via the rtloader's
pareent process path.

This change adds an ability to configure this value so that when
a library like `multiprocessing` it can point to the correct interpreter
path and work as intended.

Signed-off-by: Srdjan Grubor <sgnn7@sgnn7.org>

* [rtloader] Fixup some of the issues with the demo program

The demo program seems like it was abandoned to some extent as the paths
did not work and the callback signature was out of date. This change
allows it to be buildable and possibly runnable though the demo code has
an unrelated segmentation fault when run from Darwin that would need
futher looking into.

Signed-off-by: Srdjan Grubor <sgnn7@sgnn7.org>

* [rtloader] Launch rtloader with inferred Python interpreter's path

Now that RTLoader can accept the path to the Python interpreter as a
parameter, this change sets this value to the inferred one from the
ldflags and the environment.

Signed-off-by: Srdjan Grubor <sgnn7@sgnn7.org>

* [rtloader] Use upstream workaround for setting Python bin path on Windows

Due to an [upstream bug](https://bugs.python.org/issue34725) in how
Windows handles `Py_SetProgramName`, we have to use the suggested
workaround and also use `_Py_SetProgramFullPath` for now to make this
code work on that platform.

Since this is a workaround API in CPython, it is depricated as of
Python 3.11 and will be removed at some later version. Before it gets
removed, we should convert our Python 3 initialization code to the
more-appropriate `PyConfig API`
(https://docs.python.org/3.11/c-api/init_config.html#c.PyConfig).

* [ci] Fix issues with upgrading appveyor image

With the newer image that appveyor uses, some tests are running slower
and failing. This change fixes some tests to be more resiliant to this
change.

* [rtloader] Add `allow_python_path_heuristics_failure` flag to config

This flag can now decide if we will continue loading the rtloader or not
in cases where we can't reliably deduce the Python interpreter's full path.