[invoke] Refactor & fix release.create-rc and release.finish tasks (#8471)
Refactors most of the logic of release.create-rc and release.finish to deduplicate shared code.
Makes the tasks more future-proof by decoupling Agent 6 and Agent 7 parts, and making it easier to add future Agent versions.
Adds support for patch version bumps in release.create-rc with the --patch-version, and modifies the version fetching logic to only consider versions that match the Agent's minor version (for repositories which follow the Agent's version scheme).

Removes the optional version arguments that could be passed to release.create-rc, which made the code more complex and were seldom used (we can still add them back if we really think we need them).

Fixes a bug in release.finish where the NPM driver values could be wrong (the task would copy the contents of the last stable release instead of the last RC because we were erased said RCs first).