[profiling] expose more profiling options (#8498)
* [profiling] remove unused apiKey parameter

dd-trace-go/profiler does not require an api key, because it always
sends to the trace agent (unless WithAgentlessUpload option is used,
which our code does not).

* [profiling] expose mutex and block options

Disable mutex profiling by default, and make it configurable instead.

Add configuration options to process and system-probe agents to
configure mutex and block profiling.