[logs] Add support for new intake api (#8555)
* [logs] Add support for new intake api

Add support for new version 2 intake API used for logs and events.
New API supports additional parameters to differentiate requests: a
track type, which is part of a url, and an optional `dd-protocol`
header.

Each component that interacts that interacts with the intake, was
assigned an individual type; this is passed via various routes into the
endpoint construction.

Whether new API is used or not is controlled by a new hidden
`use_v2_api` configuratio option (available, like other options, for
each agent component separately).

Additional endpoints inherit the api version setting from the parent
endpoint, but it can be overridden if necessary.

* [logs] use correct protocol for serverless agent

* [security-agent] use correct track for intake

Security agent interacts with two different intake endpoints, and needs
to use a specific track type with each.

* [epforwarder] update track type for dbm metrics

* [logs] use net/url to build the destination url