[pkg/util] Add `hostname_file` configuration option (#8684)
In some environments, auto-detection of the hostname is not adequate and
environment variables cannot be used to set the value. In such cases, the
file on the host can also be used provide a file with the appropriate
value via `hostname_file` option.

Since getting the hostname via a file is pretty analogous to the rest of
the providers, it was made into a `HostnameProvider` and `options` were
added to the interface (and this provider may not be the only one that
could use parameterization). Ideally, we will expand this code in the
future to make the configuration file be a provider as well as there's no
need to treat the config as a bespoke provider.

Incidental to this PR some tests were added and minor formatting changes
made.