[Tagger] Expire tags for stopped ECS EC2 tasks (#8758)
This issue was introduced in #6624 and identified and fixed for a
separate collector in #6780. The same story happened here: the ECS
collector did not have its own expiration logic (even though from a
cursory inspection it would look like so from the usage of
`taggerutils.Expire`), resulting in runaway memory usage over time for
nodes executing a lot of scheduled tasks, for instance.

Besides fixing this issue, this commit also changes all appropriate
collectors to use a slightly refactored version of `taggerutils.Expire`,
centralizing a little bit of behavior.