[AD] Ignore containers stopped too long ago (#8637)
The logs agent stores a "registry.json" file with pointers to positions
inside log files of known containers, in order for a restart of the
agent not to re-send all logs from the very beginning, and instead start
at the latest position. More relevant to the changes in this commit,
this registry is cleaned of entries older than 23 hours. Unfortunately,
this makes a freshly started agent to re-send longs for containers that
were stopped longer than 23 hours ago.

To prevent this, we're introducing a `container_exclude_stopped_age`
config option to make autodiscovery ignore containers that were stopped
a certain time ago (by default, DefaultAuditorTTL minus 1 hour).