[Docker] Consistenlty report tags for stopped containers (#8776)
As stopped containers are purged from the tagger, after some 5 minutes
stopped containers will move from having all of their tags to having
none, probably messing with people's dashboards. Since a stopped
containers is guaranteed to eventually have no tags in the tagger, it is
better to treat them as we would excluded containers, and report a
synthetic set of tags that we can construct from basic container
information.