[ckey] re-introduce context key optimization (#8822)
* Revert "Revert "[aggregator/ckey] implement a faster hash to reduce CPU and RAM usage (#8529)" (#8678)"

This reverts commit 5845856cb3eeb64a54402e0341c792df4fa7f62d.

* [ckey] make Generate accept TagsBuilder

TagsBuilder is mutable slice wrapper, and this will allow us to
de-duplicate tags in it at the same as generating the context key.

* [ckey] de-duplicate tags during ckey generation

Shift tags around in the tags builder and truncate it at the end.

* [ckey] add fallback for len(tags) > 512

* [ckey] check for hash collisions

Skip tags if their hashes and string values are the same, instead of
just checking hashes.

For bruteforce method, seen and tags arrays are filled at the same time,
so adding a check is straightforward.

For hashset method, a new array seenIdx is added, and we store the
index of the tag that corresponds to the current hashset slot.

This also handles tags which hash to zero correctly.

* [ckey] use arrays to reduce bounds checks

Using arrays instead of dynamic slices, lets go drop some of the bounds
checks.

* [ckey] use smaller index array

We only store indexes up to 512, and using a smaller type means less
bytes need to be moved to initialize `seenIdx`.

* [ckey] size hashset according to tags

Initialize a part of the hashset, big enough to process incoming tags,
but not larger. Has most of positive effect when number of tags <<
hashSetSize, which makes hashset method faster than bruteforce earler.

* [ckey] add comment about hash set size

* go fmt