[kitchen] Fix kitchen logs in Gitlab artifacts (#8870)
Any time a kitchen command is run, the log files are overwritten. Due to this, since #8821, the log files we save as Gitlab artifacts only contain logs about the last operation done (the kitchen destroy). This adds the --no-log-overwrite option to all kitchen commands to save the logs of all kitchen commands.