[aggregator/ckey] tool to validate that current hash algorithm isn't creating collisions (#8526)
* [aggregator/ckey] tool to validate that current hash algorithm isn't creating collisions

The main role of the KeyGenerator we use in the aggregator is to generate hash
to represent metrics and to map them. These hashes can't collide or we could merge
different metrics as the same one, which would result to inconsistent and wrong
values.

This commit is adding a test tool running the current aggregator hash implementation
against a _lot_ of randomly generated metrics.
These random metrics and tags are only UUIDs, meaning it is not comparing against
realistic values, but should result in a valid test anyway.

It has also been validated intenally against real traffic.

* [aggregator/ckey] linter compliance on the test python script.