[packaging] Add datadog-signing-keys dependency to the deb packages (#8869)
Adds a Recommends dependency on datadog-signing-keys to the datadog-agent, datadog-iot-agent and datadog-dogstatsd deb packages.
Updates the kitchen test setup to add the datadog-signing-keys package to the testing repository, and add a kitchen test to check that the package is installed.
Updates the install script to automatically install the.datadog-signing-keys package.