[tagger] pre-hash tags (#9115)
* [ckey] move tag hashing to TagsBuilder

Hash tags as they are added to the TagsBuilder, and use these hashes
when generating context key.

Side benefit, SortUniq() for TagsBuilder() no longer allocates anything.

* [tagger/local] cache hash values for tags

Keep tags in the local tagger pre-hashed, so that we don't have to hash
the same strings over and over when generating context key.

* [tagger/local] extract tagstore package

* [tagger/local] cleanup after tagstore move

* [tagger/replay] use same tagstore as tagger/local

* Add release note

* [metrics] move EnrichTags into tagger

This code is shared between metrics, events and service checks.

* [tagger] split off hashing out of TagsBuilder

Split tag hashing function out of TagsBuilder into a new type,
HashingTagsBuilder. This avoids computing tag hashes for events and
service checks.

* [ckey] fix tags deduplication after merging main

* [tagger] use immutable hashed tags container

Split hashed tags builder into mutable (HashingTagsBuilder) and immutable
(HashedTags) variants, to prevent accidental modification of tags stored
in the taggers.

* [tagger] move nil out to caller sites

Instead of checking for nil receiver inside a method, check at the
caller site and skip the call if the value is nil.

* [ckey] small refactor

Keep slice copies local to each if arm that uses them.

* [util/tags_builder] make better use of embedding

Keep common methods on the base type; subtype specific methods can be
provided directly by the subtype.

* [util/tags_builder] restore tests for TagsBuilder

* [util/tags_builder] remove duplicate methods

Get and Hashes are provided by the base type already.

* Fix comment

* Add comment about tags layout in the tagstore

* Sync doc comments for Subscribe

subscriber.Subscribe had the most complete one, copy it to other locations.

* Fix typo

Co-authored-by: Dustin J. Mitchell <dustin.mitchell@datadoghq.com>

* [tagger/tagstore] update method name

* gofmt

* [tags_builder] improve ctor

Remove NewHashingTagsBuilderFromSlice as a potential footgun, and
replace with a constructor that copies tags instead of re-using the
slice, NewHashingTagsBuilderWithTags.

Replace the only use outside of tests with a persistent instance.

* [tagger] fix data race

Copy HashedTags out of entityTags instead of returning a pointer to a
field.

Co-authored-by: Dustin J. Mitchell <dustin.mitchell@datadoghq.com>