Lock loadedConfigs while in use (#8714)
* Lock loadedConfigs while in use

Since AD could modify the map of loaded configs at any time, it's
important that callers examining that list do so while the lock is held.
This is more efficient than copying the list.

* Remove AC.getConfigsForTemplate

This isn't required, as the only place it's used is in a racy fashion to
get the list of configs that `removeConfigsForTemplate` will remove.
That function now returns precisely the items it removed.

* Remove AC.getConfigsForService

This isn't required, as the only place it's used is in a racy fashion to
get the list of configs that `removeConfigsForService` will remove.
That function now returns precisely the items it removed.

* Add AllLoadedConfigs returning a new slice