[tooling] Add `dep_why` helper script to `dep_tree_resolver` tooling (#9308)
* [tooling] Improved docs for `dep_tree_resolver` tool

We needed a bit more exact caveats on this toos about not removing
dependencies from the output that are explicitly defined but end
up not being used by the parent module (and thus the agent).

* [tooling] Add `dep_why.py` tool to help identify dependency links

This tool automates a common use case of the dep_tree_resolver tool
output by scanning the generated dependency tree and printing out any
tree branches that lead to the specified module name. This is useful in
identifying the reasons why a particular module is being used and
through which hierarchy of dependencies.

Co-authored-by: Pablo Baeyens <pablo.baeyens@datadoghq.com>