[AD] Do not re-create existing container services (#9466)
This was a regression introduced with the workloadmeta changes. The
previously existing check for an existing container service was not
effective since it was checking for services with an ID in a different
format than the one that's used now. This fixes the check and adds a new
unit test to guard against such a regression.