[AD] Remove services for containers removed from a pod (#9528)
When `podContainers` was introduced, it tracked containers during the
entire lifetime of the pod, removing them when the pod itself went away.
This did not account for restarted containers: they would get new IDs,
but their services would not be removed until the pod was removed too.

This presented a problem for pods in CrashLoopBackoff: new services
would be generated all the time, consuming memory both in autodiscovery
with the container services, but also the tagger, as it would not allow
the gone entities to be removed from the store, as they kept being
requested.

Now, after each pod service is re-evaluated, we remove any services for
containers that were not seen in that iteration.