[tagger/workloadmeta] Use container image from pod spec
Similar to work done in c6bbe0bc2d10ac57073ffc307d3f90dce8d94bf0, we
need to expose the image name as defined by the user, instead of the one
exposed by the container runtime. This did not seem to be a problem for
k8s+docker environments, but once we added the containerd collector, we
noticed that in k8s+containerd environments, any image with an implicit
registry (say "nginx") would be reported by with the resolved registry
(so "docker.io/library/nginx"), so `image_name` tags would change and
possibly break dashboards.

Now, we have a distinction between the image as defined by the user (in
OrchestratorContainer.Image), and the actual image that the name
resolves to (in Container.Image).