[workloadmeta/containerd] Don't error out on ignoreEvent
This prevents the containerd workloadmeta collector from not starting at
all if ignoreEvent returns an error (which we've seen happen in our
environments whenever the connection to the daemon breaks, or when a
container has a reference to an image that does not exist). Since the
error returned here wasn't marked as retriable, the workloadmeta would
give up as soon as the first issue occurred, and data would not be
available.